# Conflicts:
#	playgrounds/ivtExt/src/main/java/playground/clruch/ScenarioServer.java

+++ playgrounds/ivtExt/src/main/java/playground/sebhoerl/avtaxi/framework/AVConfigGroup.java
@@ -28,7 +28,7 @@ public class AVConfigGroup extends ReflectiveConfigGroup {
 	}
 
 	@StringGetter(PARALLEL_ROUTERS)
-	public long getParallelRouters() {
+	public double getParallelRouters() {
 		return parallelRouters;
 	}
